crash fixes
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@583 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ Bundle.h
@@ -75,9 +75,9 @@ public:
 
 	typedef QueueItem* Ptr;
 	typedef unordered_map<CID, string> CIDStringList;
-	typedef unordered_map<UserPtr, uint8_t, User::Hash> UserIntMap;
+	typedef unordered_map<UserPtr, uint16_t, User::Hash> UserIntMap;
 	typedef unordered_map<TTHValue, string> FinishedItemMap;
-	typedef pair<HintedUser, uint8_t> UserRunningPair;
+	typedef pair<HintedUser, uint16_t> UserRunningPair;
 	typedef vector<UserRunningPair> SourceIntList;
 	typedef unordered_map<string, QueueItemList> DirMap;
